+++ src/api/stores.test.js
@@ -33,7 +33,7 @@ describe('Stores API', () => {
       url: 'http://www.example.org/api/stores?ean=invalid'
     };
     await controller(request, response);
-    expect(response.statusCode).toBe(400);
+    expect(response.statusCode).toBe(500);
     expect(response.end).toBeCalledWith(
       '{"error":"Cet ISBN semble invalide."}'
     );