+++ src/components/Header.js
@@ -17,7 +17,7 @@ export default function Header(props) {
             target="_blank"
             rel="noopener noreferrer"
           >
-            0.1
+            0.2
           </a>
         </span>
       </div>