+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.2.0 (2019-08-14)
+
+* Prevent error response to be cached
+* Improved handling JSON parsing error
+* Improved handling 403 error in stores API
+* Upgrade dependencies
+
 ## 0.1.0 (2019-02-15)
 
 * First release